@@ -30,11 +30,9 @@
   "dependencies": {},
   "devDependencies": {
     "cypress": "^3.2.0",
-    "eslint": "^4.16.0",
-    "eslint-config-airbnb": "^16.1.0",
+    "eslint": "^5.16.0",
+    "eslint-config-airbnb-base": "^13.1.0",
     "eslint-plugin-import": "^2.17.2",
-    "eslint-plugin-jsx-a11y": "^6.0.3",
-    "eslint-plugin-react": "^7.12.4",
     "http-server": "^0.11.1",
     "husky": "^1.3.1",
     "parcel": "^1.12.3"
