@@ -1,5 +1,7 @@
-const getRecurrPw = require('../../test-utils/v0.2.0-generator-for-test')
-const getRandom = require('../../test-utils/get-random')
+/* eslint-disable no-undef */
+
+import getRecurrPw from '../../test-utils/v0.2.0-generator-for-test'
+import { getRandom } from '../../test-utils/get-random'
 
 describe('Compatibility with the Classical Password Generator', () => {
   let password
@@ -16,7 +18,7 @@ describe('Compatibility with the Classical Password Generator', () => {
     cy.get('#password').should('have.value', '9Kb6vH')
   })
   it('generating password of length 6 with the standard charset', () => {
-    let data = ['service', 'salt', 6]
+    const data = ['service', 'salt', 6]
     getRecurrPw(...data, pw => {
       password = pw
     })
@@ -29,7 +31,7 @@ describe('Compatibility with the Classical Password Generator', () => {
     cy.get('#password').should('have.value', password)
   })
   it('generating password of length 64 with the standard charset', () => {
-    let data = ['service', 'salt', 64]
+    const data = ['service', 'salt', 64]
     getRecurrPw(...data, pw => {
       password = pw
     })
@@ -42,7 +44,7 @@ describe('Compatibility with the Classical Password Generator', () => {
     cy.get('#password').should('have.value', password)
   })
   it('generating password of big random length with the standard charset', () => {
-    let data = [
+    const data = [
       getRandom.string(1, 32), // service
       getRandom.string(1, 32), // masterpassword (salt)
       getRandom.int(6, 64), // password length
@@ -59,7 +61,7 @@ describe('Compatibility with the Classical Password Generator', () => {
     cy.get('#password').should('have.value', password)
   })
   it('generating password of moderate random length with the standard charset', () => {
-    let data = [
+    const data = [
       getRandom.string(1, 12), // service
       getRandom.string(1, 8), // masterpassword (salt)
       getRandom.int(6, 10), // password length
@@ -78,7 +80,6 @@ describe('Compatibility with the Classical Password Generator', () => {
 })
 
 describe('Modern Password Generator', () => {
-  let password
   beforeEach(() => {
     cy.visit('/')
   })
